XPRT - Keep a ref for hooked events
The epoll loop needs a ref on the XPRT, so that it stays alive while in
epoll().  Previously, the ref was taken and dropped across individual
events.  Instead, take a ref when hooking, and keep it until it's
unhooked.  This means up to 2 refs for an XPRT (one for send and one for
receive).

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>