Take a XPRT ref when hooking events
The epoll loop needs a ref on the XPRT, so that it stays alive while in
epoll().  svc_rqst_rearm_events_locked() takes a ref, but the initial
svc_rqst_hook_events() does not.  This means that, if the XPRT times
out, the cleanup will drop a ref, and the epoll loop will drop a ref,
causing one to many refs to be dropped.  This leaves the client
potentially with a dangling ref.

Take the ref for the epoll loop when initially hooking the events.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>